requirements.txt
asn1crypto==0.24.0
configparser==4.0.2
contextlib2==0.6.0.post1
cryptography==2.1.4
enum34==1.1.6
idna==2.6
importlib-metadata==2.1.3
ipaddress==1.0.17
kafka-python==2.0.2
keyring==10.6.0
keyrings.alt==3.0
numpy==1.16.6
pandas==0.24.2
pathlib2==2.3.7.post1
psycopg2==2.7.4
pycrypto==2.6.1
pygobject==3.26.1
PyMySQL==0.8.0
pyspark==3.1.2
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
pyxdg==0.25
scandir==1.10.0
SecretStorage==2.3.1
six==1.16.0
SQLAlchemy==1.4.37
typing==3.10.0.0
zipp==1.2.0
